Address review comments
The newly introduced Compilation Server message type
VM_getStringUTF8UnabbreviatedLength is never acutally used.  This change
removes it along with the VM_getStringUTF8Length message type, which is
also never used.

This change also adjusts the maximum length for a UTF-8 encoding that
can be returned by getStringUTF8Length from 2^31-1 to 2^31-2.  That
ensures that any code that then uses that length to allocate buffer to
contain the encoded String with a NUL terminator will not overflow a
32-bit signed integer representation for the length plus the NUL byte.

+++ runtime/compiler/env/VMJ9.h
@@ -1144,7 +1144,7 @@ public:
    /**
     * \brief Returns the number of UTF-8 encoded bytes needed to represent a Java String object.
     *        The number of bytes needed to UTF-8 encode the String is representable as
-    *        a \c uintptr_t, in general, but this method returns a length of type \c int32_t.
+    *        a \c uint64_t, in general, but this method returns a length of type \c int32_t.
     *        If the length might exceed the range of \c int32_t, use
     *        \ref getStringUTF8UnabbreviatedLength instead.
     *
@@ -1162,7 +1162,7 @@ public:
     * \return The number of UTF-8 encoded bytes needed to represent the String
     */
    virtual uint64_t  getStringUTF8UnabbreviatedLength(uintptr_t objectPointer);
-   virtual char     *getStringUTF8      (uintptr_t objectPointer, char *buffer, int32_t bufferSize);
+   virtual char     *getStringUTF8(uintptr_t objectPointer, char *buffer, uintptr_t bufferSize);
 
    virtual uint32_t getVarHandleHandleTableOffset(TR::Compilation *);
